+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Changed / Fixed
 - Fixed missing text domains
+- Sidebar on blog index
 
 ## [v0.1-r20170325](https://github.com/ppfeufer/eve-online-wordpress-theme/releases/tag/v0.1-r20170325) - 2017-03-25
 [Full Changelog](https://github.com/ppfeufer/eve-online-wordpress-theme/compare/v0.1-r20170307...v0.1-r20170325)